returned to old model of just one file for scene
